Fix size check in STOC_REPLAY and replay read, write (#3083)

+++ gframe/duelclient.h
@@ -21,8 +21,8 @@ public:
 	static void ClientRead(bufferevent* bev, void* ctx);
 	static void ClientEvent(bufferevent* bev, short events, void* ctx);
 	static int ClientThread();
-	static void HandleSTOCPacketLan(unsigned char* data, int len);
-	static bool ClientAnalyze(unsigned char* msg, int len);
+	static void HandleSTOCPacketLan(unsigned char* data, size_t len);
+	static bool ClientAnalyze(unsigned char* msg, size_t len);
 	static void SwapField();
 	static void SetResponseI(int32_t respI);
 	static void SetResponseB(void* respB, size_t len);
